admin: move admin server into its own directory
